# evmrs: lazy load TxContext