This PR bumps the evmc dependency when feature custom-evmc is enabled. The [new version](https://github.com/LorenzSchueler/evmc/commit/3d75575daa32e3563488361ffce84edf17109c74) no longer loads the TxContext eagerly but only when needed.

This improves performance of short running benchmarks by up to 30%.
```
                   │ 2024-12-04T15:51#50f8631#20-main/evmrs#performance │ 2024-12-04T17:38#cbb69de#20/evmrs#performance │
                   │                       sec/op                       │        sec/op          vs base                │
StaticOverhead/1/                                           2.602µ ± 0%             1.830µ ± 0%  -29.69% (p=0.000 n=20)
Inc/1/                                                      5.283µ ± 0%             4.240µ ± 0%  -19.74% (p=0.000 n=20)
Inc/10/                                                     5.314µ ± 0%             4.267µ ± 0%  -19.70% (p=0.000 n=20)
Fib/1/                                                      4.611µ ± 0%             3.698µ ± 0%  -19.80% (p=0.000 n=20)
Fib/5/                                                      14.39µ ± 0%             12.77µ ± 1%  -11.29% (p=0.000 n=20)
Fib/10/                                                     129.9µ ± 0%             119.9µ ± 0%   -7.65% (p=0.000 n=20)
Fib/15/                                                     1.224m ± 0%             1.156m ± 0%   -5.55% (p=0.000 n=20)
Fib/20/                                                     13.14m ± 0%             12.46m ± 0%   -5.18% (p=0.000 n=20)
Sha3/1/                                                     2.858µ ± 0%             2.139µ ± 0%  -25.13% (p=0.000 n=20)
Sha3/10/                                                    4.252µ ± 0%             3.526µ ± 0%  -17.09% (p=0.000 n=20)
Sha3/100/                                                   17.89µ ± 2%             16.55µ ± 0%   -7.50% (p=0.000 n=20)
Sha3/1000/                                                  188.3µ ± 0%             185.2µ ± 0%   -1.69% (p=0.000 n=20)
Arith/1/                                                    5.843µ ± 0%             5.048µ ± 0%  -13.61% (p=0.000 n=20)
Arith/10/                                                   11.29µ ± 0%             10.65µ ± 1%   -5.69% (p=0.000 n=20)
Arith/100/                                                  68.15µ ± 0%             68.99µ ± 0%   +1.24% (p=0.000 n=20)
Arith/280/                                                  199.9µ ± 0%             197.7µ ± 0%   -1.10% (p=0.000 n=20)
Memory/1/                                                   8.150µ ± 1%             7.012µ ± 1%  -13.97% (p=0.000 n=20)
Memory/10/                                                  13.40µ ± 0%             12.36µ ± 1%   -7.77% (p=0.000 n=20)
Memory/100/                                                 66.65µ ± 0%             66.80µ ± 1%   +0.24% (p=0.046 n=20)
Memory/1000/                                                574.2µ ± 0%             587.7µ ± 0%   +2.34% (p=0.000 n=20)
Memory/10000/                                               5.402m ± 0%             5.529m ± 0%   +2.35% (p=0.000 n=20)
Analysis/jumpdest/                                          2.545µ ± 0%             1.776µ ± 1%  -30.22% (p=0.000 n=20)
Analysis/stop/                                              2.536µ ± 1%             1.788µ ± 0%  -29.50% (p=0.000 n=20)
Analysis/push1/                                             2.544µ ± 0%             1.778µ ± 1%  -30.10% (p=0.000 n=20)
Analysis/push32/                                            2.546µ ± 0%             1.779µ ± 1%  -30.13% (p=0.000 n=20)
geomean                                                     26.41µ                  22.77µ       -13.78%
```

Additionally the paths that trigger rust ci workflows where adjusted to match those for go because before this PR did not trigger them.